Rename the package name from 'test' to 'e2e' for better naming convension
Tested: make e2e-test
