+++ incubator/hnc/test/e2e/delete_crd_test.go
@@ -1,4 +1,4 @@
-package test
+package e2e
 
 import (
 	"os"
