+++ incubator/hnc/test/e2e/e2e_test.go
@@ -1,4 +1,4 @@
-package test
+package e2e
 
 import (
 	"errors"
